Fixed a couple bugs
The option checking functions actually correctly set the options now by
returning them and having the result be used in a setf.

The shebang stuff now works. Before it didn't actually set the lines to
write to the correct thing.

Fixed thing where it would try to use the characters of the shebang line
as elements of the lines to write to the file.

Added blank line between shebang line and license.